+++ payment/src/main/java/net/optile/payment/ui/page/PaymentListPresenter.java
@@ -21,6 +21,7 @@ import net.optile.payment.form.Operation;
 import net.optile.payment.localization.Localization;
 import net.optile.payment.model.Interaction;
 import net.optile.payment.model.InteractionCode;
+import net.optile.payment.model.InteractionReason;
 import net.optile.payment.model.ListResult;
 import net.optile.payment.model.Redirect;
 import net.optile.payment.ui.PaymentResult;
@@ -137,7 +138,7 @@ final class PaymentListPresenter implements PaymentSessionListener, Localization
             networkService.setPresenter(this);
             preparePayment();
         } catch (PaymentException e) {
-            closeWithErrorCode(e.getMessage(), e.error);
+            closeWithErrorMessage(e.error);
         }
     }
 
@@ -146,12 +147,7 @@ final class PaymentListPresenter implements PaymentSessionListener, Localization
      */
     @Override
     public void onPaymentSessionError(Throwable cause) {
-        PaymentResult result = PaymentResult.fromThrowable(cause);
-        if (result.hasError()) {
-            handleLoadingError(result);
-        } else {
-            closeWithCanceledCode(result);
-        }
+        handleLoadingError(cause);
     }
 
     /**
@@ -164,11 +160,11 @@ final class PaymentListPresenter implements PaymentSessionListener, Localization
 
         switch (interaction.getCode()) {
             case InteractionCode.PROCEED:
-                handleLoadSessionOk(session);
+                handleLoadPaymentSessionOk(session);
                 break;
             default:
                 PaymentResult result = new PaymentResult(listResult.getResultInfo(), interaction);
-                closeWithCanceledCode(result);
+                closeWithErrorMessage(result);
         }
     }
 
@@ -191,15 +187,10 @@ final class PaymentListPresenter implements PaymentSessionListener, Localization
      */
     @Override
     public void onLocalizationError(Throwable cause) {
-        PaymentResult result = PaymentResult.fromThrowable(cause);
-        if (result.hasError()) {
-            handleLoadingError(result);
-        } else {
-            closeWithCanceledCode(result);
-        }
+        handleLoadingError(cause);
     }
 
-    private void handleLoadSessionOk(PaymentSession session) {
+    private void handleLoadPaymentSessionOk(PaymentSession session) {
         this.operation = null;
         this.session = session;
         loadLocalizations(session);
@@ -210,17 +201,16 @@ final class PaymentListPresenter implements PaymentSessionListener, Localization
         localizationService.loadLocalizations(context, session);
     }
 
-    private void handleLoadingError(PaymentResult result) {
-        PaymentError error = result.getPaymentError();
-        if (error.isError(PaymentError.CONN_ERROR)) {
-            handleLoadingConnError(result);
+    private void handleLoadingError(Throwable cause) {
+        PaymentError error = PaymentError.fromThrowable(cause);
+        if (error.isNetworkFailure()) {
+            handleLoadingNetworkFailure(error);
         } else {
-            closeWithErrorCode(result);
+            closeWithErrorMessage(error);
         }
     }
 
-    private void handleLoadingConnError(PaymentResult result) {
-        view.setPaymentResult(PaymentUI.RESULT_CODE_ERROR, result);
+    private void handleLoadingNetworkFailure(final PaymentError error) {
         view.showConnectionDialog(new ThemedDialogListener() {
             @Override
             public void onButtonClicked(ThemedDialogFragment dialog, int which) {
@@ -233,13 +223,13 @@ final class PaymentListPresenter implements PaymentSessionListener, Localization
                         }
                         break;
                     default:
-                        view.close();
+                        closeWithCanceledCode(error);
                 }
             }
 
             @Override
             public void onDismissed(ThemedDialogFragment dialog) {
-                view.close();
+                closeWithCanceledCode(error);
             }
         });
     }
@@ -254,9 +244,8 @@ final class PaymentListPresenter implements PaymentSessionListener, Localization
 
     private void handleActivityResult(ActivityResult activityResult) {
         if (this.session == null) {
-            String resultInfo = "Missing cached session in PaymentListPresenter";
-            PaymentError error = new PaymentError(PaymentError.INTERNAL_ERROR, resultInfo);
-            closeWithErrorCode(resultInfo, error);
+            PaymentError error = new PaymentError("Missing cached PaymentSession in PaymentListPresenter");
+            closeWithErrorMessage(error);
             return;
         }
         PaymentResult result = activityResult.paymentResult;
@@ -288,9 +277,6 @@ final class PaymentListPresenter implements PaymentSessionListener, Localization
             case PaymentUI.RESULT_CODE_CANCELED:
                 handlePreparePaymentCanceled(result);
                 break;
-            case PaymentUI.RESULT_CODE_ERROR:
-                handlePreparePaymentError(result);
-                break;
         }
     }
 
@@ -311,6 +297,10 @@ final class PaymentListPresenter implements PaymentSessionListener, Localization
     }
 
     private void handlePreparePaymentCanceled(PaymentResult result) {
+        if (result.hasNetworkFailureError()) {
+            handlePrepareNetworkFailure(result);
+            return;
+        }
         this.operation = null;
         Interaction interaction = result.getInteraction();
         if (interaction == null) {
@@ -327,17 +317,11 @@ final class PaymentListPresenter implements PaymentSessionListener, Localization
                 showPaymentSessionWithMessage(result);
                 break;
             default:
-                closeWithCanceledCode(result);
+                closeWithErrorMessage(result);
         }
     }
 
-    private void handlePreparePaymentError(PaymentResult result) {
-        this.operation = null;
-        if (!result.getPaymentError().isError(PaymentError.CONN_ERROR)) {
-            closeWithErrorCode(result);
-            return;
-        }
-        view.setPaymentResult(PaymentUI.RESULT_CODE_ERROR, result);
+    private void handlePrepareNetworkFailure(PaymentResult result) {
         view.showConnectionDialog(new ThemedDialogListener() {
             @Override
             public void onButtonClicked(ThemedDialogFragment dialog, int which) {
@@ -346,13 +330,13 @@ final class PaymentListPresenter implements PaymentSessionListener, Localization
                         preparePayment();
                         break;
                     default:
-                        view.close();
+                        closeWithCanceledCode(result);
                 }
             }
 
             @Override
             public void onDismissed(ThemedDialogFragment dialog) {
-                view.close();
+                closeWithCanceledCode(result);
             }
         });
     }
@@ -369,13 +353,14 @@ final class PaymentListPresenter implements PaymentSessionListener, Localization
             case PaymentUI.RESULT_CODE_CANCELED:
                 handleProcessPaymentCanceled(result);
                 break;
-            case PaymentUI.RESULT_CODE_ERROR:
-                handleProcessPaymentError(result);
-                break;
         }
     }
 
     private void handleProcessPaymentCanceled(PaymentResult result) {
+        if (result.hasNetworkFailureError()) {
+            handleProcessNetworkFailure(result);
+            return;
+        }
         this.operation = null;
         Interaction interaction = result.getInteraction();
         if (interaction == null) {
@@ -392,16 +377,11 @@ final class PaymentListPresenter implements PaymentSessionListener, Localization
                 showPaymentSessionWithMessage(result);
                 break;
             default:
-                closeWithCanceledCode(result);
+                closeWithErrorMessage(result);
         }
     }
 
-    private void handleProcessPaymentError(PaymentResult result) {
-        if (!result.getPaymentError().isError(PaymentError.CONN_ERROR)) {
-            closeWithErrorCode(result);
-            return;
-        }
-        view.setPaymentResult(PaymentUI.RESULT_CODE_ERROR, result);
+    private void handleProcessNetworkFailure(final PaymentResult result) {
         view.showConnectionDialog(new ThemedDialogListener() {
             @Override
             public void onButtonClicked(ThemedDialogFragment dialog, int which) {
@@ -410,13 +390,13 @@ final class PaymentListPresenter implements PaymentSessionListener, Localization
                         processPayment();
                         break;
                     default:
-                        view.close();
+                        closeWithCanceledCode(result);
                 }
             }
 
             @Override
             public void onDismissed(ThemedDialogFragment dialog) {
-                view.close();
+                closeWithCanceledCode(result);
             }
         });
     }
@@ -425,7 +405,7 @@ final class PaymentListPresenter implements PaymentSessionListener, Localization
         try {
             networkService.preparePayment(view.getActivity(), PREPAREPAYMENT_REQUEST_CODE, operation);
         } catch (PaymentException e) {
-            closeWithErrorCode(e.getMessage(), e.error);
+            closeWithErrorMessage(e.error);
         }
     }
 
@@ -433,7 +413,7 @@ final class PaymentListPresenter implements PaymentSessionListener, Localization
         try {
             networkService.processPayment(view.getActivity(), PROCESSPAYMENT_REQUEST_CODE, operation);
         } catch (PaymentException e) {
-            closeWithErrorCode(e.getMessage(), e.error);
+            closeWithErrorMessage(e.error);
         }
     }
 
@@ -512,26 +492,30 @@ final class PaymentListPresenter implements PaymentSessionListener, Localization
         view.close();
     }
 
+    private void closeWithCanceledCode(PaymentError error) {
+        PaymentResult result = PaymentResult.fromPaymentError(error);
+        closeWithCanceledCode(result);
+    }
+    
     private void closeWithCanceledCode(PaymentResult result) {
         view.setPaymentResult(PaymentUI.RESULT_CODE_CANCELED, result);
+        view.close();
+    }
 
+    private void closeWithErrorMessage(PaymentError error) {
+        closeWithErrorMessage(PaymentResult.fromPaymentError(error));
+    }
+    
+    private void closeWithErrorMessage(PaymentResult result) {
+        view.setPaymentResult(PaymentUI.RESULT_CODE_CANCELED, result);
         String msg = translateInteraction(result.getInteraction());
         if (TextUtils.isEmpty(msg)) {
             msg = Localization.translate(ERROR_DEFAULT);
         }
-        closeWithMessage(msg);
+        showMessageAndClose(msg);
     }
 
-    private void closeWithErrorCode(String resultInfo, PaymentError error) {
-        closeWithErrorCode(new PaymentResult(resultInfo, error));
-    }
-
-    private void closeWithErrorCode(PaymentResult result) {
-        view.setPaymentResult(PaymentUI.RESULT_CODE_ERROR, result);
-        closeWithMessage(Localization.translate(ERROR_DEFAULT));
-    }
-
-    private void closeWithMessage(String message) {
+    private void showMessageAndClose(String message) {
         view.showMessageDialog(message, new ThemedDialogListener() {
             @Override
             public void onButtonClicked(ThemedDialogFragment dialog, int which) {
@@ -562,9 +546,5 @@ final class PaymentListPresenter implements PaymentSessionListener, Localization
             return null;
         }
         return Localization.translateInteraction(interaction);
-    }
-
-    private String getString(int resId) {
-        return view.getActivity().getString(resId);
     }
 }